requirements.txt
streamlit>=1.36.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
python-dateutil>=2.8.2
reportlab>=4.0.0
pytest>=8.0.0
